+++ pkg/analyzer/lib/src/dart/resolver/binary_expression_resolver.dart
@@ -460,9 +460,7 @@ class BinaryExpressionResolver {
     var leftOperand = node.leftOperand;
 
     TypeImpl leftType;
-    if (leftOperand is AugmentedExpressionImpl) {
-      leftType = leftOperand.typeOrThrow;
-    } else if (leftOperand is ExtensionOverrideImpl) {
+     if (leftOperand is ExtensionOverrideImpl) {
       leftType = leftOperand.extendedType!;
     } else {
       leftType = leftOperand.typeOrThrow;