+++ pkg/analyzer/lib/src/test_utilities/find_node.dart
@@ -32,10 +32,6 @@ class FindNode {
 
   AssignmentExpression get singleAssignmentExpression => _single();
 
-  AugmentedExpression get singleAugmentedExpression => _single();
-
-  AugmentedInvocation get singleAugmentedInvocation => _single();
-
   AwaitExpression get singleAwaitExpression => _single();
 
   BinaryExpression get singleBinaryExpression => _single();