+++ pkg/test_runner/lib/src/configuration.dart
@@ -11,7 +11,6 @@ import 'package:smith/configuration.dart';
 import 'package:smith/smith.dart';
 
 import 'compiler_configuration.dart';
-import 'deflake_info.dart';
 import 'feature.dart';
 import 'path.dart';
 import 'repository.dart';
@@ -32,8 +31,7 @@ class TestConfiguration {
       this.selectors = const {},
       this.build = false,
       this.testList = const [],
-      this.deflakeInfoMap = const {},
-      int repeat = 1,
+      this.repeat = 1,
       this.batch = false,
       this.copyCoreDumps = false,
       this.rr = false,
@@ -76,8 +74,7 @@ class TestConfiguration {
       : packages = packages ??
             Repository.uri
                 .resolve('.dart_tool/package_config.json')
-                .toFilePath(),
-        _repeat = repeat;
+                .toFilePath();
 
   final Map<String, RegExp?> selectors;
   final Progress progress;
@@ -139,12 +136,11 @@ class TestConfiguration {
   final String? dartPrecompiledPath;
   final String? genSnapshotPath;
   final List<String>? testList;
-  final Map<String, DeflakeInfo> deflakeInfoMap;
 
   final int taskCount;
   final int shardCount;
   final int shard;
-  final int _repeat;
+  final int repeat;
 
   final int testServerPort;
   final int testServerCrossOriginPort;
@@ -225,15 +221,15 @@ class TestConfiguration {
   ///     build/ReleaseX64
   String get buildDirectory => system.outputDirectory + configurationDirectory;
 
-  int? _defaultTimeout;
+  int? _timeout;
 
   // TODO(whesse): Put non-default timeouts explicitly in configs, not this.
   /// Calculates a default timeout based on the compiler and runtime used,
   /// and the mode, architecture, etc.
-  int get defaultTimeout {
-    if (_defaultTimeout == null) {
+  int get timeout {
+    if (_timeout == null) {
       if (configuration.timeout > 0) {
-        _defaultTimeout = configuration.timeout;
+        _timeout = configuration.timeout;
       } else {
         var isReload = hotReload || hotReloadRollback;
 
@@ -245,21 +241,12 @@ class TestConfiguration {
             arch: architecture,
             system: system);
 
-        _defaultTimeout = 30 * compilerMultiplier * runtimeMultiplier;
+        _timeout = 30 * compilerMultiplier * runtimeMultiplier;
       }
     }
 
-    return _defaultTimeout!;
+    return _timeout!;
   }
-
-  /// Returns the timeout for the given test name.
-  int timeout(String name) {
-    final t = deflakeInfoMap[name]?.timeout ?? -1;
-    return t >= 0 ? t : defaultTimeout;
-  }
-
-  /// Returns the repeat count for the given test name.
-  int repeat(String name) => deflakeInfoMap[name]?.repeat ?? _repeat;
 
   List<String> get standardOptions {
     if (compiler != Compiler.dart2js) {